fixed 2 warnings with -Wsign-compare

+++ src/include/mallocMC/distributionPolicies/XMallocSIMD_impl.hpp
@@ -77,7 +77,7 @@ namespace DistributionPolicies{
 
       //all the properties must be unsigned integers > 0
       BOOST_STATIC_ASSERT(!std::numeric_limits<typename Properties::pagesize::type>::is_signed);
-      BOOST_STATIC_ASSERT(pagesize > 0);
+      BOOST_STATIC_ASSERT(static_cast<uint32>(pagesize) > 0);
 
     public:
       static const uint32 _pagesize = pagesize;
